+++ packages/atomic/src/components/common/facets/numeric-facet-common.tsx
@@ -363,7 +363,7 @@ export class NumericFacetCommon {
         isCollapsed={isCollapsed}
         headingLevel={this.props.headingLevel}
         onToggleCollapse={onToggleCollapse}
-        headerRef={headerFocus.setTarget}
+        headerRef={(el) => headerFocus.setTarget(el)}
       ></FacetHeader>
     );
   }