+++ packages/headless-react/jest.config.js
@@ -1,0 +1,12 @@
+/** @type {import('ts-jest').JestConfigWithTsJest} */
+module.exports = {
+  preset: 'ts-jest',
+  testEnvironment: 'jsdom',
+  collectCoverage: true,
+  collectCoverageFrom: ['./src/**.ts', './src/**.tsx'],
+  coverageThreshold: {
+    global: {
+      lines: 90,
+    },
+  },
+};