+++ packages/headless/src/test/mock-engine.ts
@@ -26,7 +26,7 @@ import {SearchEngine} from '../app/search-engine/search-engine';
 import {SearchThunkExtraArguments} from '../app/search-thunk-extra-arguments';
 import {CaseAssistEngine} from '../case-assist.index';
 import {CaseAssistAppState} from '../state/case-assist-app-state';
-import {ProductListingV2AppState} from '../state/commerce-app-state';
+import {CommerceAppState} from '../state/commerce-app-state';
 import {InsightAppState} from '../state/insight-app-state';
 import {ProductListingAppState} from '../state/product-listing-app-state';
 import {ProductRecommendationsAppState} from '../state/product-recommendations-app-state';
@@ -34,10 +34,10 @@ import {RecommendationAppState} from '../state/recommendation-app-state';
 import {SearchAppState} from '../state/search-app-state';
 import {validatePayloadAndThrow} from '../utils/validate-payload';
 import {buildMockCaseAssistState} from './mock-case-assist-state';
+import {buildMockCommerceState} from './mock-commerce-state';
 import {buildMockInsightAPIClient} from './mock-insight-api-client';
 import {buildMockInsightState} from './mock-insight-state';
 import {buildMockProductListingState} from './mock-product-listing-state';
-import {buildMockProductListingV2State} from './mock-product-listing-v2-state';
 import {buildMockProductRecommendationsState} from './mock-product-recommendations-state';
 import {createMockRecommendationState} from './mock-recommendation-state';
 import {buildMockSearchAPIClient} from './mock-search-api-client';
@@ -58,7 +58,7 @@ type AppState =
   | RecommendationAppState
   | ProductRecommendationsAppState
   | ProductListingAppState
-  | ProductListingV2AppState
+  | CommerceAppState
   | CaseAssistAppState
   | InsightAppState;
 
@@ -147,19 +147,13 @@ export function buildMockProductListingEngine(
 }
 
 export interface MockCommerceEngine
-  extends CommerceEngine<ProductListingV2AppState>,
+  extends CommerceEngine<CommerceAppState>,
     MockEngine {}
 
-/**
- * For internal use only.
- *
- * Returns a non-functional `CommerceEngine`.
- * To be used only for unit testing controllers, not for functional tests.
- */
 export function buildMockCommerceEngine(
-  config: Partial<CommerceEngine<ProductListingV2AppState>> = {}
+  config: Partial<CommerceEngine<CommerceAppState>> = {}
 ): MockCommerceEngine {
-  return buildMockCoreEngine(config, buildMockProductListingV2State);
+  return buildMockCoreEngine(config, buildMockCommerceState);
 }
 
 export interface MockCaseAssistEngine