+++ packages/samples/headless-ssr/src/app/generic/components/search-page.tsx
@@ -6,7 +6,8 @@ import {
   hydrateInitialState,
 } from '@/src/app/generic/common/engine';
 import {useEffect, useState} from 'react';
-import {HydrationMetadata} from './hydration-metadata';
+import {HydrationMetadata} from '../../common/hydration-metadata';
+import {useSyncSearchParameters} from '../hooks/search-parameters';
 import {ResultList} from './result-list';
 import {SearchBox} from './search-box';
 
@@ -16,13 +17,25 @@ export default function SearchPage({ssrState}: {ssrState: SearchSSRState}) {
   );
 
   useEffect(() => {
-    hydrateInitialState(ssrState).then(({engine, controllers}) => {
+    hydrateInitialState({
+      searchFulfilledAction: ssrState.searchFulfilledAction,
+      controllers: {
+        searchParameters: {
+          initialState: ssrState.controllers.searchParameters.state,
+        },
+      },
+    }).then(({engine, controllers}) => {
       setCSRResult({engine, controllers});
     });
   }, [ssrState]);
+
+  useSyncSearchParameters({
+    ssrState: ssrState.controllers.searchParameters.state,
+    controller: csrResult?.controllers.searchParameters,
+  });
+
   return (
     <>
-      <HydrationMetadata ssrState={ssrState} csrResult={csrResult} />
       <SearchBox
         ssrState={ssrState.controllers.searchBox.state}
         controller={csrResult?.controllers.searchBox}
@@ -31,6 +44,7 @@ export default function SearchPage({ssrState}: {ssrState: SearchSSRState}) {
         ssrState={ssrState.controllers.resultList.state}
         controller={csrResult?.controllers.resultList}
       />
+      <HydrationMetadata ssrState={ssrState} csrResult={csrResult} />
     </>
   );
 }